+++ conan/api/subapi/new.py
@@ -82,9 +82,9 @@ class NewAPI:
     @staticmethod
     def get_builtin_template(template_name):
         from conan.internal.api.new.basic import basic_file
+        from conan.internal.api.new.basic import flattened_file
         from conan.internal.api.new.alias_new import alias_file
         from conan.internal.api.new.cmake_exe import cmake_exe_files
-        from conan.internal.api.new.demo import demo_files
         from conan.internal.api.new.cmake_lib import cmake_lib_files
         from conan.internal.api.new.header_lib import header_only_lib_files
         from conan.internal.api.new.meson_lib import meson_lib_files
@@ -105,7 +105,6 @@ class NewAPI:
         new_templates = {"basic": basic_file,
                          "cmake_lib": cmake_lib_files,
                          "cmake_exe": cmake_exe_files,
-                         "demo": demo_files,
                          "header_lib": header_only_lib_files,
                          "meson_lib": meson_lib_files,
                          "meson_exe": meson_exe_files,
@@ -124,7 +123,7 @@ class NewAPI:
                          "local_recipes_index": local_recipes_index_files,
                          "qbs_lib": qbs_lib_files,
                          "workspace": workspace_files}
-        template_files = new_templates.get(template_name)
+        template_files = new_templates.get(template_name, flattened_file)
         return template_files
 
     def get_template(self, template_folder):
